Additions on behalf of freezing the layered texture during updating.
Summary:
Three additions on behalf of freezing the layered texture during updating,
which is done in the geometry-nodes (TexturePlane, TexturePanelStrip, etc.)

(1) Functionality to turn on/off the displaying of any texture altogether.
(2) Dirtying the tiles of images with stride is now delayed until demanded,
because dirtying them on the osg-level will break the texture freeze.
(3) The border color of the composite texture (no shading) was not computed
so far, but will now show up during the texture freeze stage after enlarging
the geometry.
(4) A fex cosmetic changes.

Test Plan:
(1) Tested in example_textureplane and via visBase::textureChannels::turnOn()
(2) Resizing an inl/crl/z-slice (Display->Position) with multiple attributes
while freezing the texture (to be committed next), and seeing that texture
does not change until all loading is finished.
(3) As in (2), but with shading disabled. Play with the undefined color and
transparency of the multiple attributes, and see that now the same border
colors show up during the texture freeze as when shading is enabled.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D993