Seam management functionality to minimize mipmapping-induced tile artifacts.
Summary:
(1) Bugfix for prominent mipmapping-induced tiling artifacts that appear when   
texel size on screen is not (about) square, because of e.g. unequal sampling    
or z-stretching. The one-sided one-texel wide seam used so far needed to        
be replaced by a double-sided seam of adaptive width, for which a suitable      
minimum value is derived automatically. Function setSeamPower(.) enables a      
manual raise. Altogether, it required major rework of all tiling algorithms.    
(2) Bugfix for double composite texture creation and flashy update on screen    
for no-shader option. Occurring since revision D993 [item (3)] that sets the    
border color of the composite texture, which triggered yet another update.      
All code dealing with the new _reInitTiling variable belongs to this.

Test Plan:
(1) Testing in od_osg with planes and 2D lines with an attribute (e.g. energy)  
at which the artifacts were most prominent. Playing with the z-stretch, and     
seeing that the artifacts are gone. Lots of testing with example_textureplane   
in osgGeo to check whether the rework did not destroy existing functionality.   
However, the LayeredTexture class contains so many degrees of freedom that      
exhaustive testing is not feasible. An occasional imperfection should show      
itself in future.                                                               
(2) Tested with example_textureplane in osgGeo. Toggle shaders on/off, which    
triggers update from update traversal. Toggle image layers, which triggers      
update from cull traversal. Always consider both these scenarios when dealing   
with update issues in LayeredTexture. No flashing update on screen anymore.     
Breakpoint at createCompositeTexture() to assure that it is created only once.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1323