Enabling the mouse pickability of well logs, and a few related bugfixes.
Summary:
(1) Added functionality to serve intersection traversals in order to make the   
well logs pickable. Since these are added to the scene on-the-fly because of    
their camera orientation dependency, this does not happen automatically.
(2) Improved BoundingSphere computation and management so that a well is only   
culled when out of sight and so that the mouse picks on a well log are never    
clipped away.                                                                   
(3) Bugfix for displaying REPEATED seismic logs, whose mathematics turned out   
not at all invariant under scene rotation. Guess this was not tested earlier.

Test Plan:
(1-2) Moving the mouse pointer over well logs and (repeated) seismic logs at    
different camera orientations and zooms, and see that log information is shown  
in the status bar of OpendTect for any position on the logs. Right now, this    
is only the case for triangles and not for lines. The latter requires OD's      
EventCatchHandler to be extended with OSG's PolytopeIntersector functionality.  
(3) The sequence of repeated seismic logs is no longer deforming when rotating  
the scene.

Reviewers: ding

Reviewed By: ding

Differential Revision: https://phabricator.dgbes.com:8443/D1615