Custom TabBoxDragger class to overrule default tab size and facilitate future customizations.
Summary:
Created custom TabBoxDragger class to set tab size of its six TabPlaneDraggers.
Because these are initialized in the constructor, osgManiputor::TabBoxDragger 
cannot be used as base class, but had to be copied instead.

Test Plan:
TabBoxDragger is used for volumes in OpendTect and now got tabs with sizes
comparable to Coin3D.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2654