Customized Translate(1D/2D)Dragger which can be set to become active only when clicked without any mod key.
Summary:
The code of the handle(.) function in Dragger.cpp offers no possibility to
become active only if the dragger is clicked without any mod key, and return
"false" otherwise. Logic of the setActivationModKeyMask(.) function is such
that one can specify which mod keys are required to become active, but not
which mod keys are forbidden to become active. Unclear whether this is a bug,
a design flaw or missing functionality. Added a setInactivationModKeyMask(.)
function to the derived classes. Did not choose to add it to a customized
version of the Dragger base class instead, because that would force the
customized derived classes to copy the code from their osg-implementations,
which screws up maintainability.

Test Plan:
Bug occurred in OpendTect when trying to remove fault (stickset) knots by
control-clicking. Control-double clicking was used as workaround so far.
Problem was solved after having setInactivationModKeyMask(.) called by the
draggers in OpendTect.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2907