osgGeo::computerBoundingBoxVisitor
Summary:
The new class and All changes in osgGeo are used for re-implement the computation of boundingbox for osg::computeBoundingVistor.
This solves the viewAll function in od. However, for the time being, we can not remove updateCalculatedNearFar in Cull_visitor. it will 
leads to objects gone when doing rotation. So this needs to be further investigated.

Test Plan: Compile and tested with set updateCalculatedNearFar working fine.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2950