Bugfix: new bounding box computation did not account for marker size.
Summary:
New bounding box computation did not account for marker size. This causes that
marker sets with only one marker are not displayed at all. It also causes that
markers located on the boundary of the bounding box are only pickable on part
of their surface. Beware that the constant screen size of the markers makes
their size in the 3D scene change dynamically, and therefore the bounding box
as well.

Test Plan: Tested on faultstick sets and picksets in OpendTect.

Reviewers: ding

Reviewed By: ding

Differential Revision: https://phabricator.dgbes.com:8443/D2960