Reversed order in which shader code of coltab process loads its channels.
Summary:
Better to reverse order in which shader code of coltab process loads its
channels, so that the texture coordinates of the original signal are finally
used to derive the mip-value. It would only make a difference if the power
channel images are scaled differently, which is unlikely but not forbidden.

Test Plan:
Only tested whether color table textures in OpendTect, which make use of power
channels, did not change appearance.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3545