Fixed the bug due to setting depth buffer issue.
Summary:
The original changes are good for catching intersection information but missed the line geometry which re-generates a bug when set fill color 
as none, the line geometry can not be displayed anymore. and also after set none, set fill color back, it would not be worked.

Test Plan: Compiles and Tests fine

Reviewers: jaap

Reviewed By: jaap

Differential Revision: https://phabricator.dgbes.com:8443/D3668