Added isProjectionMatrixInited function.
Summary:
The purpose of this function is to check if the camera projection matrix is set before calling viewAll function, if it is not set, we will set a default projection Matrix
via updateCamera function.
Because sometimes the view hasn't start to do traverse, the viewAll function has been called already. Especial when we create a new scene.

Test Plan: Compiles and Test fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D4497