Partial reverse of r427/D5177 issue (3) about contour line-like artifacts in zoomed-out low-frequent textures (PORTED FROM od TO od5.0 UNCHANGED)
Summary:
Contour line-like artifacts in zoomed-out low-frequent textures at some
hardware are caused by a failing GLSL sqrt-function for values close to 0.0.
Now EMPERICAL value 0.0001 is used as threshold. Minimizing its effect on the
result by moving a factor 255.0 inside the sqrt-expression turned out to
reintroduce the artifact. Looks like the failure has a linear relationship
with the magnitude of the sqrt-expression.

Test Plan: Patched on Ding's computer to conclude that the artifact has gone.

Reviewers: ding

Reviewed By: ding

Differential Revision: https://phabricator.dgbes.com:8443/D5783