Bugfix: Textures were not displayed on virtual machines (Mantis 3484)
Summary:
Textures were not displayed on virtual machines. Issue (1) of r405/D4561 has
to be reverted. It just disables that what it was supposed to solve. Mea culpa. After a discussion in June 2014 on od4.6 virtual machine problems, I verified
whether our osgGeo code could deal with it. Apparently, I overlooked the checks
already taking care of this, and added new ones in the wrong place, shadowing
the old ones. On top of that, I wrote: "Not (yet) tested whether it actually
solves virtual machines problems, but adding these shader availability checks
will not hurt". How bad can it be.

Test Plan:
Tested on a virtual windows machine, which shows textures on inlines and
horizons again after reverting issue (1) of r405/D4561.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6220